+++ CHANGELOG.md
@@ -9,11 +9,16 @@ __BEGIN_UNRELEASED__
 ## [Unreleased]
 ### Added
 ### Changed
+- Updated the way events are sent to the Heap backend to allow for first-class support of React Native as a library.  See <TODO: upgrade guide link> for details.
+- Upgraded the native Android Heap SDK to v1.3.0.
+- Upgraded vendored iOS Heap library to 6.5.1.
 - Updated higher-order components to use display name conventions (described [here](https://reactjs.org/docs/higher-order-components.html#convention-wrap-the-display-name-for-easy-debugging)).
 
 ### Deprecated
 ### Removed
 ### Fixed
+- Fixed React Navigation autocapture when [manual screen tracking](https://reactnavigation.org/docs/en/screen-tracking.html) is also used.
+
 ### Security
 __END_UNRELEASED__